requirements.txt
# LLM Orchestrator Bot Dependencies
anthropic>=0.66.0
slack-bolt>=1.24.0
mcp[cli]>=1.13.1
python-dotenv>=1.1.1
aiohttp>=3.12.15
certifi>=2024.12.14
boto3>=1.40.26
